requirements.txt
apipkg==1.5
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
execnet==1.7.1
idna==2.10
iniconfig==1.1.1
packaging==20.7
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
pytest-assume==2.4.2
pytest-forked==1.3.0
pytest-rerunfailures==9.1.1
pytest-xdist==2.1.0
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
toml==0.10.2
urllib3==1.26.2

selenium~=3.141.0